+++ apps/explorer/lib/explorer/chain/address.ex
@@ -19,7 +19,6 @@ defmodule Explorer.Chain.Address do
     Hash,
     InternalTransaction,
     SmartContract,
-    SmartContractAdditionalSource,
     Token,
     Transaction,
     Wei,
@@ -79,7 +78,6 @@ defmodule Explorer.Chain.Address do
              :contracts_creation_internal_transaction,
              :contracts_creation_transaction,
              :names,
-             :smart_contract_additional_sources,
              :tags
            ]}
 
@@ -92,7 +90,6 @@ defmodule Explorer.Chain.Address do
              :contracts_creation_internal_transaction,
              :contracts_creation_transaction,
              :names,
-             :smart_contract_additional_sources,
              :tags
            ]}
 
@@ -128,7 +125,6 @@ defmodule Explorer.Chain.Address do
 
     has_many(:names, Address.Name, foreign_key: :address_hash)
     has_many(:decompiled_smart_contracts, DecompiledSmartContract, foreign_key: :address_hash)
-    has_many(:smart_contract_additional_sources, SmartContractAdditionalSource, foreign_key: :address_hash)
     has_many(:withdrawals, Withdrawal, foreign_key: :address_hash)
     has_many(:tags, AddressTag, foreign_key: :id)
 
@@ -348,11 +344,11 @@ defmodule Explorer.Chain.Address do
   @doc """
   Checks if given address is smart-contract
   """
-  @spec is_smart_contract(any()) :: boolean() | nil
-  def is_smart_contract(%__MODULE__{contract_code: nil}), do: false
-  def is_smart_contract(%__MODULE__{contract_code: _}), do: true
-  def is_smart_contract(%NotLoaded{}), do: nil
-  def is_smart_contract(_), do: false
+  @spec smart_contract?(any()) :: boolean() | nil
+  def smart_contract?(%__MODULE__{contract_code: nil}), do: false
+  def smart_contract?(%__MODULE__{contract_code: _}), do: true
+  def smart_contract?(%NotLoaded{}), do: nil
+  def smart_contract?(_), do: false
 
   defp get_addresses(options) do
     accounts_with_n = fetch_top_addresses(options)
@@ -389,5 +385,59 @@ defmodule Explorer.Chain.Address do
         (address.fetched_coin_balance == ^coin_balance and address.hash > ^hash) or
           address.fetched_coin_balance < ^coin_balance
     )
+  end
+
+  @doc """
+  Checks if an `t:Explorer.Chain.Address.t/0` with the given `hash` exists.
+
+  Returns `:ok` if found
+
+      iex> {:ok, %Explorer.Chain.Address{hash: hash}} = Explorer.Chain.create_address(
+      ...>   %{hash: "0x5aaeb6053f3e94c9b9a09f33669435e7ef1beaed"}
+      ...> )
+      iex> Explorer.Address.check_address_exists(hash)
+      :ok
+
+  Returns `:not_found` if not found
+
+      iex> {:ok, hash} = Explorer.Chain.string_to_address_hash("0x5aaeb6053f3e94c9b9a09f33669435e7ef1beaed")
+      iex> Explorer.Address.check_address_exists(hash)
+      :not_found
+
+  """
+  @spec check_address_exists(Hash.Address.t(), [Chain.api?()]) :: :ok | :not_found
+  def check_address_exists(address_hash, options \\ []) do
+    address_hash
+    |> address_exists?(options)
+    |> Chain.boolean_to_check_result()
+  end
+
+  @doc """
+  Checks if an `t:Explorer.Chain.Address.t/0` with the given `hash` exists.
+
+  Returns `true` if found
+
+      iex> {:ok, %Explorer.Chain.Address{hash: hash}} = Explorer.Chain.create_address(
+      ...>   %{hash: "0x5aaeb6053f3e94c9b9a09f33669435e7ef1beaed"}
+      ...> )
+      iex> Explorer.Chain.Address.address_exists?(hash)
+      true
+
+  Returns `false` if not found
+
+      iex> {:ok, hash} = Explorer.Chain.string_to_address_hash("0x5aaeb6053f3e94c9b9a09f33669435e7ef1beaed")
+      iex> Explorer.Chain.Address.address_exists?(hash)
+      false
+
+  """
+  @spec address_exists?(Hash.Address.t(), [Chain.api?()]) :: boolean()
+  def address_exists?(address_hash, options \\ []) do
+    query =
+      from(
+        address in Address,
+        where: address.hash == ^address_hash
+      )
+
+    Chain.select_repo(options).exists?(query)
   end
 end